Add create course tag api for screeners
